Update validation test layout
