+++ tests/validation/test_feature_api.py
@@ -1,0 +1,80 @@
+import unittest
+
+import numpy as np
+
+from ballet.eng.base import BaseTransformer
+from ballet.eng.misc import IdentityTransformer
+from ballet.compat import SimpleImputer
+from ballet.feature import Feature
+from ballet.validation.feature_api import (
+    CanDeepcopyCheck,
+    CanTransformCheck,
+    HasCorrectInputTypeCheck,
+    HasCorrectOutputDimensionsCheck,
+    validate)
+
+from .util import SampleDataMixin
+from ..util import FragileTransformer
+
+
+class ProjectStructureTest(SampleDataMixin, unittest.TestCase):
+
+    def test_good_feature(self):
+        feature = Feature(
+            input='size',
+            transformer=SimpleImputer(),
+        )
+
+        result, failures = validate(feature, self.X, self.y)
+        self.assertTrue(result)
+        self.assertEqual(len(failures), 0)
+
+    def test_bad_feature_input(self):
+        # bad input
+        feature = Feature(
+            input=3,
+            transformer=SimpleImputer(),
+        )
+        result, failures = validate(feature, self.X, self.y)
+        self.assertFalse(result)
+        self.assertIn(HasCorrectInputTypeCheck.__name__, failures)
+
+    def test_bad_feature_transform_errors(self):
+        # transformer throws errors
+        feature = Feature(
+            input='size',
+            transformer=FragileTransformer(
+                (lambda x: True, ), (RuntimeError, ))
+        )
+        result, failures = validate(feature, self.X, self.y)
+        self.assertFalse(result)
+        self.assertIn(CanTransformCheck.__name__, failures)
+
+    def test_bad_feature_wrong_transform_length(self):
+        class _WrongLengthTransformer(BaseTransformer):
+            def transform(self, X, **transform_kwargs):
+                new_shape = list(X.shape)
+                new_shape[0] += 1
+                output = np.arange(np.prod(new_shape)).reshape(new_shape)
+                return output
+
+        # doesn't return correct length
+        feature = Feature(
+            input='size',
+            transformer=_WrongLengthTransformer(),
+        )
+        result, failures = validate(feature, self.X, self.y)
+        self.assertFalse(result)
+        self.assertIn(HasCorrectOutputDimensionsCheck.__name__, failures)
+
+    def test_bad_feature_deepcopy_fails(self):
+        class _CopyFailsTransformer(IdentityTransformer):
+            def __deepcopy__(self):
+                raise RuntimeError
+        feature = Feature(
+            input='size',
+            transformer=_CopyFailsTransformer(),
+        )
+        result, failures = validate(feature, self.X, self.y)
+        self.assertFalse(result)
+        self.assertIn(CanDeepcopyCheck.__name__, failures)
