+++ tests/validation/test_project_structure.py
@@ -2,184 +2,17 @@ import unittest
 from textwrap import dedent
 from unittest.mock import patch
 
-import numpy as np
-import pandas as pd
-from funcy import contextmanager
 
-from ballet.compat import SimpleImputer
-from ballet.eng.base import BaseTransformer
-from ballet.eng.misc import IdentityTransformer
-from ballet.feature import Feature
 from ballet.util.ci import TravisPullRequestBuildDiffer
 from ballet.util.git import get_diff_str_from_commits
-from ballet.validation import (
-    FeatureApiValidator, FileChangeValidator, SingleFeatureApiValidator)
+from ballet.validation.project_structure import (
+    FileChangeValidator)
 
-from .util import (
-    FragileTransformer, make_mock_commit, make_mock_commits, mock_repo)
-
-
-@contextmanager
-def mock_project(path_content):
-    with mock_repo() as repo:
-        for path, content in path_content:
-            make_mock_commit(repo, path=path, content=content)
-        yield repo
-
-
-@contextmanager
-def null_file_change_validator(pr_num):
-    with mock_repo() as repo:
-        commit_range = 'HEAD^..HEAD'
-        contrib_module_path = None
-        X = None
-        y = None
-
-        travis_env_vars = {
-            'TRAVIS_BUILD_DIR': repo.working_tree_dir,
-            'TRAVIS_PULL_REQUEST': str(pr_num),
-            'TRAVIS_COMMIT_RANGE': commit_range,
-        }
-
-        with patch.dict('os.environ', travis_env_vars, clear=True):
-            yield FileChangeValidator(
-                repo, pr_num, contrib_module_path, X, y)
-
-
-@contextmanager
-def mock_file_change_validator(
-    path_content, pr_num, contrib_module_path, X, y
-):
-    """FileChangeValidator for mock repo and mock project content
-
-    Args:
-        path_content: iterable of (relative path, file content)
-    """
-    with mock_project(path_content) as repo:
-        travis_build_dir = repo.working_tree_dir
-        travis_pull_request = str(pr_num)
-        travis_commit_range = get_diff_str_from_commits(
-            repo.head.commit.parents[0], repo.head.commit)
-
-        travis_env_vars = {
-            'TRAVIS_BUILD_DIR': travis_build_dir,
-            'TRAVIS_PULL_REQUEST': travis_pull_request,
-            'TRAVIS_COMMIT_RANGE': travis_commit_range,
-        }
-
-        with patch.dict('os.environ', travis_env_vars, clear=True):
-            yield FileChangeValidator(
-                repo, pr_num, contrib_module_path, X, y)
-
-
-@contextmanager
-def mock_feature_api_validator(
-    path_content, pr_num, contrib_module_path, X, y
-):
-    """FileChangeValidator for mock repo and mock project content
-
-    Args:
-        path_content: iterable of (relative path, file content)
-    """
-    with mock_project(path_content) as repo:
-        travis_build_dir = repo.working_tree_dir
-        travis_pull_request = str(pr_num)
-        travis_commit_range = get_diff_str_from_commits(
-            repo.head.commit.parents[0], repo.head.commit)
-
-        travis_env_vars = {
-            'TRAVIS_BUILD_DIR': travis_build_dir,
-            'TRAVIS_PULL_REQUEST': travis_pull_request,
-            'TRAVIS_COMMIT_RANGE': travis_commit_range,
-        }
-
-        with patch.dict('os.environ', travis_env_vars, clear=True):
-            yield FeatureApiValidator(
-                repo, pr_num, contrib_module_path, X, y)
-
-
-class SampleDataMixin:
-    def setUp(self):
-        self.df = pd.DataFrame(
-            data={
-                'country': ['USA', 'USA', 'Canada', 'Japan'],
-                'year': [2001, 2002, 2001, 2002],
-                'size': [np.nan, -11, 12, 0.0],
-                'strength': [18, 110, np.nan, 101],
-                'happy': [False, True, False, False]
-            }
-        ).set_index(['country', 'year'])
-        self.X = self.df[['size', 'strength']]
-        self.y = self.df[['happy']]
-        super().setUp()
-
-
-class SingleFeatureApiValidatorTest(SampleDataMixin, unittest.TestCase):
-
-    def test_good_feature(self):
-        feature = Feature(
-            input='size',
-            transformer=SimpleImputer(),
-        )
-
-        validator = SingleFeatureApiValidator(self.X, self.y)
-        result, failures = validator.validate(feature)
-        self.assertTrue(result)
-        self.assertEqual(len(failures), 0)
-
-    def test_bad_feature_input(self):
-        # bad input
-        feature = Feature(
-            input=3,
-            transformer=SimpleImputer(),
-        )
-        validator = SingleFeatureApiValidator(self.X, self.y)
-        result, failures = validator.validate(feature)
-        self.assertFalse(result)
-        self.assertIn('has_correct_input_type', failures)
-
-    def test_bad_feature_transform_errors(self):
-        # transformer throws errors
-        feature = Feature(
-            input='size',
-            transformer=FragileTransformer(
-                (lambda x: True, ), (RuntimeError, ))
-        )
-        validator = SingleFeatureApiValidator(self.X, self.y)
-        result, failures = validator.validate(feature)
-        self.assertFalse(result)
-        self.assertIn('can_transform', failures)
-
-    def test_bad_feature_wrong_transform_length(self):
-        class _WrongLengthTransformer(BaseTransformer):
-            def transform(self, X, **transform_kwargs):
-                new_shape = list(X.shape)
-                new_shape[0] += 1
-                output = np.arange(np.prod(new_shape)).reshape(new_shape)
-                return output
-
-        # doesn't return correct length
-        feature = Feature(
-            input='size',
-            transformer=_WrongLengthTransformer(),
-        )
-        validator = SingleFeatureApiValidator(self.X, self.y)
-        result, failures = validator.validate(feature)
-        self.assertFalse(result)
-        self.assertIn('has_correct_output_dimensions', failures)
-
-    def test_bad_feature_deepcopy_fails(self):
-        class _CopyFailsTransformer(IdentityTransformer):
-            def __deepcopy__(self):
-                raise RuntimeError
-        feature = Feature(
-            input='size',
-            transformer=_CopyFailsTransformer(),
-        )
-        validator = SingleFeatureApiValidator(self.X, self.y)
-        result, failures = validator.validate(feature)
-        self.assertFalse(result)
-        self.assertIn('can_deepcopy', failures)
+from .util import (mock_feature_api_validator,
+                   mock_file_change_validator,
+                   null_file_change_validator,
+                   SampleDataMixin)
+from ..util import make_mock_commits, mock_repo
 
 
 class ProjectStructureValidatorTest(SampleDataMixin, unittest.TestCase):
